Parsing termine, balbutiements traitement

+++ src/utils.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int				ft_strlen(const char *s)
+#include <stddef.h>
+
+size_t			ft_strlen(const char *s)
 {
 	const char	*s1;
